requirements.txt
customtkinter
openai-whisper
pyttsx3
pydub
